Feat(DPU): update to support hydration and streaming race conditions
